Cassandra tests messed up?
