+++ atlasdb-tests-shared/src/main/java/com/palantir/atlasdb/sweep/AbstractBackgroundSweeperIntegrationTest.java
@@ -16,18 +16,15 @@
 package com.palantir.atlasdb.sweep;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.primitives.Ints;
 import com.palantir.atlasdb.AtlasDbConstants;
-import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.RangeRequest;
 import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.api.Value;
 import com.palantir.atlasdb.keyvalue.impl.SweepStatsKeyValueService;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.SweepStrategy;
 import com.palantir.atlasdb.schema.generated.SweepTableFactory;
@@ -48,19 +45,15 @@ import com.palantir.atlasdb.transaction.service.TransactionServices;
 import com.palantir.atlasdb.util.MetricsManager;
 import com.palantir.atlasdb.util.MetricsManagers;
 import com.palantir.common.base.ClosableIterator;
-import com.palantir.common.streams.KeyedStream;
 import com.palantir.lock.SingleLockService;
 import com.palantir.timestamp.InMemoryTimestampService;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.function.BooleanSupplier;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -79,7 +72,7 @@ public abstract class AbstractBackgroundSweeperIntegrationTest {
             .candidateBatchSize(15)
             .maxCellTsPairsToExamine(1000)
             .build();
-    private TransactionService txService;
+    protected TransactionService txService;
     SpecificTableSweeper specificTableSweeper;
     AdjustableSweepBatchConfigSource sweepBatchConfigSource;
     DeterministicTrueSupplier skipCellVersion = new DeterministicTrueSupplier();
@@ -168,141 +161,6 @@ public abstract class AbstractBackgroundSweeperIntegrationTest {
                 .isTrue();
     }
 
-    @Test
-    public void previouslyConservativeSweepsEverythingWhenNothingIsSkipped() {
-        skipCellVersion.setPeriod(Integer.MAX_VALUE);
-        createTable(TABLE_1, SweepStrategy.THOROUGH);
-        putManyCells(TABLE_1, 100, 101);
-        putManyCells(TABLE_1, 103, 104);
-        putManyCells(TABLE_1, 107, 109);
-
-        SweeperService sweeperService = new SweeperServiceImpl(specificTableSweeper, sweepBatchConfigSource);
-        sweeperService.sweepPreviouslyConservativeNowThoroughTable(
-                TABLE_1.getQualifiedName(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty());
-
-        verifyTableSwept(TABLE_1, 58, false);
-    }
-
-    @Test
-    public void previouslyConservativeErasesExistingSentinelsInThoroughTable() {
-        skipCellVersion.setPeriod(Integer.MAX_VALUE);
-        createTable(TABLE_1, SweepStrategy.THOROUGH);
-
-        Map<Cell, byte[]> sentinelWrites = IntStream.range(0, 100)
-                .mapToObj(PtBytes::toBytes)
-                .map(bytes -> Cell.create(bytes, bytes))
-                .collect(Collectors.toMap(cell -> cell, _ignore -> PtBytes.EMPTY_BYTE_ARRAY));
-        kvs.put(TABLE_1, sentinelWrites, Value.INVALID_VALUE_TIMESTAMP);
-        Map<Cell, Long> readMap = KeyedStream.stream(sentinelWrites)
-                .map(_ignore -> Long.MAX_VALUE)
-                .collectToMap();
-        assertThat(kvs.get(TABLE_1, readMap)).hasSize(100);
-
-        SweeperService sweeperService = new SweeperServiceImpl(specificTableSweeper, sweepBatchConfigSource);
-        sweeperService.sweepPreviouslyConservativeNowThoroughTable(
-                TABLE_1.getQualifiedName(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty());
-
-        assertThat(kvs.get(TABLE_1, readMap)).isEmpty();
-    }
-
-    @Test
-    public void previouslyConservativeThrowsIfTableIsStillConservativelySwept() {
-        createTable(TABLE_1, SweepStrategy.CONSERVATIVE);
-
-        SweeperService sweeperService = new SweeperServiceImpl(specificTableSweeper, sweepBatchConfigSource);
-        assertThatThrownBy(() -> sweeperService.sweepPreviouslyConservativeNowThoroughTable(
-                        TABLE_1.getQualifiedName(),
-                        Optional.empty(),
-                        Optional.empty(),
-                        Optional.empty(),
-                        Optional.empty(),
-                        Optional.empty()))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessage("it is not safe to run this type of sweep on conservatively swept tables");
-    }
-
-    @Test
-    public void previouslyConservativeRespectsSkipPeriodWhenErasingSentinels() {
-        skipCellVersion.setPeriod(4);
-        createTable(TABLE_1, SweepStrategy.THOROUGH);
-
-        Map<Cell, byte[]> sentinelWrites = IntStream.range(0, 100)
-                .mapToObj(PtBytes::toBytes)
-                .map(bytes -> Cell.create(bytes, bytes))
-                .collect(Collectors.toMap(cell -> cell, _ignore -> PtBytes.EMPTY_BYTE_ARRAY));
-        kvs.put(TABLE_1, sentinelWrites, Value.INVALID_VALUE_TIMESTAMP);
-        Map<Cell, Long> readMap = KeyedStream.stream(sentinelWrites)
-                .map(_ignore -> Long.MAX_VALUE)
-                .collectToMap();
-        assertThat(kvs.get(TABLE_1, readMap)).hasSize(100);
-
-        SweeperService sweeperService = new SweeperServiceImpl(specificTableSweeper, sweepBatchConfigSource);
-        sweeperService.sweepPreviouslyConservativeNowThoroughTable(
-                TABLE_1.getQualifiedName(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty());
-
-        // Impl specific, documenting here -- exact last row will be swept twice, so the sentinel will be erased on
-        // second pass-through; this is fine, not worth the risk of modifying behaviour
-        assertThat(kvs.get(TABLE_1, readMap)).hasSize(24);
-    }
-
-    @Test
-    public void previouslyConservativeRespectsSkipPeriodWhenSweepingNormally() {
-        skipCellVersion.setPeriod(9);
-        createTable(TABLE_1, SweepStrategy.THOROUGH);
-
-        Map<Cell, byte[]> writes = KeyedStream.of(IntStream.range(0, 1000).boxed())
-                .mapKeys(PtBytes::toBytes)
-                .mapKeys(bytes -> Cell.create(bytes, bytes))
-                .map(count -> count < 500 ? PtBytes.toBytes(count) : PtBytes.EMPTY_BYTE_ARRAY)
-                .collectToMap();
-        kvs.put(TABLE_1, writes, 100L);
-        txService.putUnlessExists(100L, 101L);
-        kvs.put(TABLE_1, writes, 103L);
-        txService.putUnlessExists(103L, 104L);
-
-        Map<Cell, Long> readMap =
-                KeyedStream.stream(writes).map(_ignore -> 102L).collectToMap();
-        assertThat(kvs.get(TABLE_1, readMap)).hasSize(1000);
-
-        SweeperService sweeperService = new SweeperServiceImpl(specificTableSweeper, sweepBatchConfigSource);
-        sweeperService.sweepPreviouslyConservativeNowThoroughTable(
-                TABLE_1.getQualifiedName(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty());
-
-        // deletes all but a ninth of the entries at the lower timestamp
-        assertThat(kvs.get(TABLE_1, readMap)).hasSize(111);
-        Map<Cell, Long> readsAtMaxTs =
-                KeyedStream.stream(readMap).map(_ignore -> Long.MAX_VALUE).collectToMap();
-        // none of the cells at lower timestamp became visible
-        assertThat(kvs.get(TABLE_1, readsAtMaxTs).values().stream()
-                        .map(Value::getTimestamp)
-                        .filter(timestamp -> timestamp == 100L)
-                        .collect(Collectors.toList()))
-                .hasSize(0);
-        // 500 cannot be completely removed because they are not deletes, 111 of the remaining 500 have either of the
-        // two versions skipped from sweep, which results in a version here
-        assertThat(kvs.get(TABLE_1, readsAtMaxTs)).hasSize(500 + 111);
-    }
-
     protected abstract KeyValueService getKeyValueService();
 
     void verifyTableSwept(TableReference tableRef, int expectedCells, boolean conservative) {
@@ -365,7 +223,7 @@ public abstract class AbstractBackgroundSweeperIntegrationTest {
         return 150L;
     }
 
-    private static final class DeterministicTrueSupplier implements BooleanSupplier {
+    protected static final class DeterministicTrueSupplier implements BooleanSupplier {
         private int truePeriod = 1;
         private long count = 0;
